add hookable for erc20, delete default impl for erc20
