+++ crates/components/token/erc20/src/lib.rs
@@ -3,6 +3,7 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 
 mod erc20;
+mod erc20_hookable;
 mod extensions;
 mod module;
 
@@ -14,15 +15,23 @@ pub use metis_lang::{
 
 pub use module::Data;
 
-pub use erc20::{
+pub use erc20_hookable::{
     Error,
     EventEmit,
-    Impl,
     Result,
+    Impl,
 };
+
+pub mod default {
+    pub use super::erc20::Impl;
+}
+
+pub mod hookable {
+    pub use super::erc20_hookable::Impl;
+}
 
 pub mod burnable {
     pub use super::extensions::burnable::Impl;
 }
 
-impl<E: Env, T: Storage<E, Data<E>> + EventEmit<E>> Impl<E> for T {}
+// impl<E: Env, T: Storage<E, Data<E>> + EventEmit<E>> Impl<E> for T {}
